ci: More comprehensive CI workflow, run on PRs
CI now runs cargo check, clippy, fmt and test (with default features,
no features, and all features), and ensures the cargo lock file is up to
date when running each.

The fail-fast mode is disabled to ensure we can still run all checks
in the case that one fails so that we can get a more comprehensive
breakdown of what needs tweaking on PRs, rather than just seeing the
first job that fails.

The cargo-toml-lint cmd is run separately at the moment as the necessary
command itself is a bit unwieldy and doesn't fit into the cargo cmd
matrix.