feat: Make BytecodeMapped generic over its bytecode slice type
Previously, `BytecodeMapped` could either:

- Consume and map an owned `Vec<u8>`.
- Construct the inner `bytecode` `Vec<u8>` lazily as operations are
  pushed to it.

This PR adds a third option which is to borrow and map an existing slice
of bytes by making the inner `bytecode` type generic. This allows for
constructing `BytecodeMapped` around a `&[u8]`, `Arc<[u8]>`, or any
other types that derefence to a slice of bytes.

This should help us to avoid the need to use `exec_bytecode_iter`
downstream in the essential server as seen here:

https://github.com/essential-contributions/essential-server/blob/80be250cc413df70f301c34570ce66a682f9e40a/crates/essential/src/solution.rs#L299-L305

Instead, we can use `exec_bytecode` and make use of the existing slice.

Closes #64.