feat: Enable `#![no-std]` support for *-asm crates
Small tweaks to the `*-asm` crates to make them `#![no_std]` compatible.
This just adds some flexibility, and makes sure that CI will notify us
if we lose the option by adding something from `std`.

I also intended to update `types`, but it's a little more ambiguous
on how to go about it as there are a few ways to go about dealing with
`Vec` (i.e. introduing alloc, switching to ArrayVec, etc).