fix: Update/clarify `IntentAddress` docs on how to calculate `set` address (#115)
* fix: Update/clarify `IntentAddress` docs on how to construct `set` addr

It's tricky to enforce the correctness of this through types or by
construction while having public fields and while `essential-hash`
is downstream from `essential-types`, but I think documenting the
expectation clearly like this is probably enough.

We don't actually ever calculate this in `essential-base` as most tests
use placeholder `IntentAddress`es, but we do in the `essential-server`
when we store intents. I'll open up a sister PR at `essential-server`
before landing this.

Closes #114.

* feat: Add mod of shorthand `intent_set_addr` functions to essential-hash

Provide helpers for common scenarios when calculating the intent set
address.