refactor: Rm unused `rayon` crate
Previously the constraint VM offered a `check_constraints` fn that
checked constraints in parallel using rayon.

We never used this in `essential-check`, and instead had a custom
parallel implementation in check itself in order to get rayon working
nicely with tokio.

Since switching to the program DAG, we no longer execute constraints
separately in parallel, and instead use tokio tasks to execute programs
asynchronously.

We may re-add parallelism of the leaf node programs (constraints) in the
future, but until then we don't make use of `rayon` otherwise.