Merge the constraint and state-read ASM and VM crates into `essential-asm`, `essential-vm` (#229)
* feat: Consolidate constraint and state read ASM in asm-gen, asm-spec

* refactor: Merge constraint-asm and state-asm into new essential-asm crate

* refactor: WIP merge constraint and state-read VMs into essential-vm

* refactor: WIP continued merging VMs

* refactor: Move repeats test into vm tests. Rm old constraint-vm crate.

* fix: Add missing error types to VM. Remove old error and lib mods.

* refactor: Update essential-check for merge of VMs and ASM

* fix: Run cargo fmt

* fix: Add missing trace_op_res, shared between async and sync execution

* fix: Run cargo fmt

* fix: More removals of state read terminology throughout repo

* fix: ASM doc tweaks

* fix: Run cargo fmt

* refactor: Refactor ASM operations, make StateRead subgroup, remove Constraint

* refactor: Update essential-vm and check for ASM refactor

Refactors the `constraint` module into a `sync` module that better
reflects the distinction between the set of ops originally referred to
as `Constraint` ops. Updates related error and result types accordingly.

* fix: Address typo found in review

* refactor: Rm unused `rayon` crate

Previously the constraint VM offered a `check_constraints` fn that
checked constraints in parallel using rayon.

We never used this in `essential-check`, and instead had a custom
parallel implementation in check itself in order to get rayon working
nicely with tokio.

Since switching to the program DAG, we no longer execute constraints
separately in parallel, and instead use tokio tasks to execute programs
asynchronously.

We may re-add parallelism of the leaf node programs (constraints) in the
future, but until then we don't make use of `rayon` otherwise.

* fix: Re-add accidentally deleted part of asm test

* fix: Rm unused OpSyncError variant NextPcOverflow

* fix: Correct stack input order for StoreRange op

* feat: Add simple tests for memory ops that test actual operations

* fix: Correct jumbled tests in asm crate

* fix: Make imports consistent in vm pred module

* fix: Type inference error in asm test only occurring with all-features enabled